# tests/FindMyCat.UnitTests/Services/Hologram/hologram-openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Hologram REST API v1",
...
    "version": "1"
  },
  "servers": [
    {
      "url": "https://dashboard.hologram.io/api/1",
      "description": "Hologram REST API v1"
    }
  ],
  "paths": {
...
    "/devices": {
      "get": {
        "tags": [
          "Devices"
        ],
        "summary": "List devices",
        "description": "Retrieve a list of devices with their detailed information including profile information (cellular links), session data, and metadata. The devices are returned sorted by creation date, with the most recent devices appearing first. Use the \u0060limit\u0060 and \u0060startafter\u0060 parameters together for cursor-based pagination through large device lists.",
        "operationId": "getDevices",
        "parameters": [
          {
            "name": "orgid",
            "in": "query",
            "description": "Filter results to this organization ID",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of results to return (max 5000)",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "example": 50
            }
          },
          {
            "name": "startafter",
            "in": "query",
            "description": "Return results after the record with this ID (cursor-based pagination)",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "slim",
            "in": "query",
            "description": "When set to 1, return a compact response with fewer fields",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "withlocation",
            "in": "query",
            "description": "When true, include last known location data in each device record (default: true)",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "states",
            "in": "query",
            "description": "Comma-separated list of device states to filter by (e.g. LIVE,PAUSED-USER,CONNECTED)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ids",
            "in": "query",
            "description": "Comma-separated list of device IDs to filter by",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tagid",
            "in": "query",
            "description": "Filter results to devices with this tag ID",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "tagids",
            "in": "query",
            "description": "Comma-separated list of tag IDs to filter by",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tagname",
            "in": "query",
            "description": "Filter results to devices with this tag name",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "query",
            "description": "Filter results to devices matching this exact name",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "imei",
            "in": "query",
            "description": "Filter results to devices with this IMEI",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "linkid",
            "in": "query",
            "description": "Filter results to devices with this cellular link ID",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "linkids",
            "in": "query",
            "description": "Comma-separated list of cellular link IDs to filter by",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "planid",
            "in": "query",
            "description": "Filter results to devices subscribed to this data plan ID",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "iccid",
            "in": "query",
            "description": "Filter results to devices with this ICCID",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "iccids",
            "in": "query",
            "description": "Comma-separated list of ICCIDs to filter by",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sim",
            "in": "query",
            "description": "Alias for iccid \u2014 filter results to devices with this ICCID",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sims",
            "in": "query",
            "description": "Alias for iccids \u2014 comma-separated list of ICCIDs to filter by",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "imsi",
            "in": "query",
            "description": "Filter results to devices with this IMSI on their cellular link",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "msisdn",
            "in": "query",
            "description": "Filter results to devices with this MSISDN on their cellular link",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "simcardid",
            "in": "query",
            "description": "Filter results to devices with this SIM card ID",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "eid",
            "in": "query",
            "description": "Filter results to devices with this EID",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "eids",
            "in": "query",
            "description": "Comma-separated list of EIDs to filter by",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "eidrange",
            "in": "query",
            "description": "Filter by EID range in \u0022start-end\u0022 format",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "showHidden",
            "in": "query",
            "description": "When true, include hidden devices in results",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "phone",
            "in": "query",
            "description": "Filter results to devices with this phone number",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tunnelable",
            "in": "query",
            "description": "Filter results to devices with this tunnelable flag value",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListDevicesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "x-mint": {
          "metadata": {
            "sidebarTitle": "List devices"
          },
          "href": "/api/1/devices/get"
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "NetworkState": {
        "type": "string",
        "enum": [
          "INACTIVE",
          "INACTIVE-TESTED",
          "FACTORY-PENDING",
          "FACTORY",
          "LIVE-PENDING",
          "LIVE",
          "FROZEN",
          "FROZEN-PENDING",
          "DEAD",
          "DEAD-PENDING",
          "PAUSED-SYS",
          "PAUSED-PENDING-SYS",
          "PAUSED-USER",
          "PAUSED-PENDING-USER",
          "TEST-ACTIVATE",
          "TEST-ACTIVATE-PENDING",
          "TEST-AUTOACTIVATE",
          "TEST-AUTOACTIVATE-PENDING",
          "TEST-HIBERNATE",
          "TEST-HIBERNATE-PENDING"
        ]
      },
      "ProfileState": {
        "type": "string",
        "enum": [
          "AVAILABLE",
          "DOWNLOAD_PENDING",
          "DOWNLOADED",
          "ENABLE_PENDING",
          "ENABLED",
          "DISABLE_PENDING",
          "DISABLED",
          "DELETE_PENDING",
          "ASSIGNED",
          "UNAVAILABLE"
        ]
      },
...
      "AsyncChangeRequestObject": {
        "description": "A non-terminal async change request record.",
        "properties": {
          "id": {
            "description": "Async change request ID",
            "type": "integer"
          },
          "parent_id": {
            "description": "Parent request ID for child requests",
            "type": "integer",
            "nullable": true
          },
          "orgid": {
            "description": "Organization ID",
            "type": "integer",
            "nullable": true
          },
          "userid": {
            "description": "User ID who initiated the request",
            "type": "integer",
            "nullable": true
          },
          "request_type": {
            "description": "Type of async change request",
            "type": "string",
            "enum": [
              "plan_change",
              "euicc_audit",
              "sim_configuration_update",
              "configure_sim",
              "sim_state_pause",
              "sim_state_unpause",
              "sim_state_activation"
            ]
          },
          "reference_type": {
            "description": "Type of the referenced object",
            "type": "string",
            "enum": [
              "device",
              "link",
              "simcard",
              "sim_configuration"
            ]
          },
          "reference_id": {
            "description": "ID of the referenced object",
            "type": "integer"
          },
          "state": {
            "description": "Current state of the request",
            "type": "string",
            "enum": [
              "initialized",
              "in_progress"
            ]
          },
          "start_time": {
            "description": "Timestamp when the request was created (UTC)",
            "type": "string"
          },
          "update_time": {
            "description": "Timestamp of the last state update (UTC)",
            "type": "string"
          },
          "metadata": {
            "description": "Request-type-specific metadata",
            "type": "object",
            "nullable": true
          }
        },
        "type": "object"
      },
      "CellularLinkInDevice": {
        "description": "A cellular link (SIM profile) associated with a device.",
        "properties": {
          "id": {
            "description": "Cellular link ID",
            "type": "integer"
          },
          "deviceid": {
            "description": "Device ID",
            "type": "integer"
          },
          "orgid": {
            "description": "Organization ID",
            "type": "integer"
          },
          "sim": {
            "description": "SIM ICCID",
            "type": "string"
          },
          "msisdn": {
            "description": "Phone number assigned to this SIM",
            "type": "string"
          },
          "imsi": {
            "description": "International Mobile Subscriber Identity",
            "type": "integer"
          },
          "state": {
            "$ref": "#/components/schemas/NetworkState"
          },
          "profile_state": {
            "$ref": "#/components/schemas/ProfileState"
          },
          "apn": {
            "description": "Access Point Name",
            "type": "string"
          },
          "zone": {
            "description": "Coverage zone ID",
            "type": "integer"
          },
          "overagelimit": {
            "description": "Data overage limit in bytes (-1 for unlimited)",
            "type": "integer"
          },
          "carrier": {
            "description": "Carrier backend ID",
            "type": "integer"
          },
          "whencreated": {
            "description": "SIM activation timestamp (UTC)",
            "type": "string"
          },
          "whenexpires": {
            "description": "SIM expiry timestamp (UTC)",
            "type": "string"
          },
          "async_change_requests": {
            "description": "Non-terminal async change requests associated with this cellular link",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AsyncChangeRequestObject"
            }
          }
        },
        "type": "object"
      },
      "DeviceObject": {
        "properties": {
          "id": {
            "description": "Unique device ID",
            "type": "integer"
          },
          "name": {
            "description": "Device name",
            "type": "string"
          },
          "orgid": {
            "description": "Organization ID",
            "type": "integer"
          },
          "type": {
            "description": "Device type",
            "type": "string"
          },
          "whencreated": {
            "description": "Device creation timestamp (UTC)",
            "type": "string"
          },
          "tags": {
            "description": "Tags associated with this device",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "links": {
            "properties": {
              "cellular": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CellularLinkInDevice"
                }
              }
            },
            "type": "object"
          },
          "eid": {
            "description": "EID of the sim card, may be null for some sim card types",
            "type": "string",
            "nullable": true
          },
          "imei": {
            "description": "IMEI of the device",
            "type": "string"
          },
          "imei_sv": {
            "description": "IMEI Software Version (IMEISV) of the device",
            "type": "string"
          },
          "tunnelable": {
            "description": "1 if the device supports SpaceBridge tunneling",
            "type": "integer",
            "enum": [
              0,
              1
            ]
          },
          "hidden": {
            "description": "1 if the device is hidden on the Dashboard",
            "type": "integer",
            "enum": [
              0,
              1
            ]
          },
          "phonenumber": {
            "description": "Phone number assigned to the device, empty string if none",
            "type": "string"
          },
          "phonenumber_cost": {
            "description": "Monthly cost of the assigned phone number",
            "type": "string"
          },
          "simcardid": {
            "description": "SIM card ID associated with the device",
            "type": "integer"
          },
          "intended_network_state": {
            "$ref": "#/components/schemas/NetworkState"
          },
          "intended_plan_id": {
            "description": "Intended plan ID for the SIM",
            "type": "integer",
            "nullable": true
          },
          "euicc_tested": {
            "description": "eUICC test status for the device",
            "type": "string"
          },
          "is_hyper": {
            "description": "True if the device uses a Hyper eUICC SIM rather than a standard UICC",
            "type": "boolean"
          },
          "model": {
            "description": "Device model derived from IMEI TAC lookup, \u0022Unknown\u0022 if not found",
            "type": "string"
          },
          "manufacturer": {
            "description": "Device manufacturer derived from IMEI TAC lookup, \u0022Unknown\u0022 if not found",
            "type": "string"
          },
          "preflight_starttime": {
            "description": "Timestamp when the preflight period started, null if not in preflight",
            "type": "string",
            "nullable": true
          },
          "preflight_scheduled_endtime": {
            "description": "Timestamp when the preflight period is scheduled to end, null if not in preflight",
            "type": "string",
            "nullable": true
          },
          "async_change_requests": {
            "description": "Non-terminal async change requests associated with this device",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AsyncChangeRequestObject"
            }
          }
        },
        "type": "object"
      },
      "ListDevicesResponse": {
        "properties": {
          "success": {
            "$ref": "#/components/schemas/SuccessProperty"
          },
          "limit": {
            "type": "integer"
          },
          "size": {
            "type": "integer"
          },
          "continues": {
            "type": "boolean"
          },
          "lastid": {
            "description": "ID of the last record returned, for cursor-based pagination",
            "type": "integer"
          },
          "links": {
            "$ref": "#/components/schemas/PaginationLinks"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DeviceObject"
            }
          }
        },
        "type": "object"
      },
...
      "PaginationLinks": {
        "properties": {
          "path": {
            "description": "The API endpoint path",
            "type": "string"
          },
          "base": {
            "description": "The API base URL",
            "type": "string"
          },
          "next": {
            "description": "URL to retrieve the next page of results",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "SuccessProperty": {
        "description": "Indicates whether the request was successful.",
        "type": "boolean",
        "example": true
      }
    },
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "http",
        "description": "HTTP Basic authentication using API keys. Set the username to \u0060apikey\u0060 and the password to your API key. You can find your API key on the Hologram Dashboard under Account Settings.",
        "scheme": "basic"
      },
      "bearerAuth": {
        "type": "http",
        "description": "Bearer token authentication using an access token.",
        "bearerFormat": "JWT",
        "scheme": "bearer"
      }
    }
  },
...
}